Doors, Object Pooling, and Player Weapons
-Added Door functionality.
a. Doors are instantiated in dungeon rooms at the location of corridors.
b. Doors will open once player enters trigger box.
c. Added functionality for doors to be locked (this will depend on whether there are enemies remain in the room)

- Object Pooling added
a. Object Pool will be used for bullets and enemy gameObjects.
b. This is to reduce garbage collection.
c. Pool size to be determined in Editor

-Player Weapons
a. Each player character will have a starting weapon that will have infinite ammo
b. Added funcitonality for player weapon firing and reloading.
c. These are called using Events
d. Weapon UI also added which subscribe to the Weapon firing and reloading events which will update the corresponding UI elements.
e. If player ammo is depleted player will not be able to fire or reload gun until new ammo is found. (Implementation pending)