Enemy Pathfinding, Enemy Spawning, Enemy Attacks, Health and Damage
-Enemy Pathfinding
a. Enemies utilize A* Pathfinding in combination with preferred path tiles.
b. Enemies will find the shortest route to the player utilizing preferred tile paths which are pre-defined on each Room prefab. The preferred paths are weighted with a value of 1 while non-preferred paths are weighted with a value of 40.
c. This approach was used in order to ensure enemies avoid obstacles in a more believable manner and to prevent the AI from getting stuck on obstacles.
d. If the player is outside the preferred path the enemy will still follow the preferred path until the last second before transitioning to non-preferred paths.

- Optimization of A* pathfinding
a. Utilizing Unity's profiler tool it was clear that the pathfinding algorithm was utilizing a large amount of CPU power the more enemies there on the screen. This was due to simultaneous calculations of all enemy AI when the player moved enough to trigger a path recalculation.
b. To combat this the AI A* calculations have been spread across multiple frames, with a target of 60fps. Each enemy will calculate their new path on a unique frame with up to 60 enemies being able to have their own unique frame. This spreads the CPU power across multiple frames reducing spiking and fps drops.
c. Further optimization could be done by implementing a binary-heap in the pathfinding algorithm.

-Enemy Spawning
a. Spawning of enemies is unique to each type of dungeon room determined in the DungeonLevelSO and RoomSO.
b. Spawning is confined to small, medium, large, and boss rooms. No spawning in corridors, entrances, or chest rooms.
c. Each Room SO has it's own list of enemies, number of enemies to spawn, and how many enemies to spawn at once.
d. This is configured in the Unity Editor in the dungeon room prefabs.

-Enemy Weapons.
a. Enemy Weapon AI takes in the player's position, distance to the player, and whether or not the player is visible.
b. The weapon AI will rotate the weapon towards the player and fire once the player is in range and if the player is visible (e.g. not behind an obstacle.)

-Health and Damage
a. Health and Damage is handled through the HealthEvent and DamageEvents.
b. If the gameObject isDamageable these events will trigger and will update the UI in the case of player.
c. If the player isRolling then isDamageable will be false during the animation, making them invulnerable and bullets to pass through them.
d. If the gameObject is marked immuneAfterHit then the gameObject will be immune to damage for a short duration. This is mainly to prevent players from being hit by multiple bullets at once and dying immediately.
d. A check is also done to see if isColliding is true, if it is then it will not cause further damage. This is to prevent one bullet hitting the same gameObject multiple times and doubling its damage by hitting multiple contact points on a gameObject's collider.
e. ContactDamage was implemented which works in two parts. ReceivingContactDamage and DealContactDamage.
f. Both enemies and players are able to DealContactDamage but only the player and destructibhle environment objects will ReceiveContactDamage.
g. If the player receives contact damage from an enemy but is still touching the enemy collider once their hit immunity is expired they will receive contact damage again until they move away from the enemy collider.
h. If the gameObject is currently affected by hit immunity they will flash red until the hit immunity has expired.